Currently working on the new monster computer.

+++ computeIDFromResult.m
@@ -57,18 +57,18 @@ function [Issues] = computeIDFromResult(Issues, solution)
     
     statetime = solutionstatestable.getIndependentColumn();
     
-%     starttime = statetime.get(0).doubleValue();
-%     endtime = statetime.get(statetime.size()-1).doubleValue();
-    starttime = statetime.get(0);
-    endtime = statetime.get(statetime.size()-1);
+    starttime = statetime.get(0).doubleValue();
+    endtime = statetime.get(statetime.size()-1).doubleValue();
+    % starttime = statetime.get(0);
+    % endtime = statetime.get(statetime.size()-1);
 
     timelength = statetime.size();
     
     for i=0:timelength-1
         temprow = solutionstatestable.getRowAtIndex(i).getAsMat();
         temprow2 = org.opensim.modeling.Vector().createFromMat(temprow);
-        sto.append(statetime.get(i), temprow2);
-%         sto.append(statetime.get(i).doubleValue(), temprow2);
+        sto.append(statetime.get(i).doubleValue(), temprow2);
+        % sto.append(statetime.get(i), temprow2);
     end
     
     % idstorage = solution.exportToStatesStorage();
